requirements.txt
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
pytz==2018.7
six==1.11.0
unicodecsv==0.14.1
